SUNSUCK

+++ include/gale/util.h
@@ -6,7 +6,7 @@ extern struct gale_dir *dot_gale,*home_dir,*sys_dir;
 void gale_init(const char *,int argc,char * const *argv);
 void gale_restart(void);
 
-void *gale_malloc(int size);
+void *gale_malloc(size_t size);
 void gale_free(void *);
 
 void *gale_memdup(const void *,int);
